# project/liftLog-trainUp/src/docs/swagger.doc.json
{"openapi":"3.0.0","info":{"tittle":"Workout Match","description":"Create, copy and shared custom workouts and exercises between users","contact":{"email":"arthurgbsf@gmail.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:3000","description":"Local Database"}],"paths":{"/users/authentication":{"post":{"sumary":"Authenticate User","description":"This route is responsible for authenticate a user.","tags":["Auth"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"},"examples":{"user":{"value":{"email":"joao@example.com","password":"123"}}}}}},"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/User"},"headers":{"Authorization":{"description":"Autohrization token","schema":{"type":"string"}}}}}},"400":{"description":"Bad Request."},"401":{"description":"Email and Password is necessary."},"404":{"description":"User not Found."}}}},"/users":{"post":{"sumary":"Create User","description":"Route to create an user","tags":["Users"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"},"examples":{"user":{"value":{"name":"Arthur Spitz","email":"arthurspitz@example.com","password":"123456"}}}}}},"responses":{"201":{"description":"User created","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/User"}}}},"401":{"description":"Name, Email and Password are required."},"404":{"description":"Email already registered"}}},"get":{"sumary":"Get all users","description":"Route to get all users","tags":["Users"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"type":"object","items":{"$ref":"#/components/schemas/User"}}}}},"400":{"description":"Bad Request."},"401":{"description":"Unauthorized."},"404":{"description":"No users registered."}}},"put":{"sumary":"Update logged user","description":"Responsible for update the logged user","tags":["Users"],"security":[{"bearerAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"},"examples":{"Update user":{"value":{"name":"Arthur Fitz","email":"arthurfitz@example.com","password":"123456"}},"Update user email":{"value":{"email":"arthurfitz@example.com"}},"Update user password":{"value":{"password":"654321"}},"Update user nameame":{"value":{"name":"Arthur Fitz"}}}}}},"responses":{"200":{"description":"User updated","content":{"application/json":{"schema":{"type":"object","items":{"$ref":"#/components/schemas/User"}}}}},"400":{"description":"Just allowed: |Name |Email |Password"},"404":{"description":"Email already registered."}}},"delete":{"sumary":"Delete logged user","description":"Responsible to delete the logged user","tags":["Users"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"The user was deleted"},"400":{"description":"User wasn't delete"}}},"/users/user":{"get":{"sumary":"Get logged user information","description":"This route is responsible for get logged user information","tags":["User"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"type":"object","items":{"$ref":"#/components/schemas/User"}}}}},"404":{"description":"User not found."}}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}